Minor improvements for models & services
