+++ src/main/java/com/leverx/javacourse/seller_rating_app/SellerRatingApplication.java
@@ -2,10 +2,8 @@ package com.leverx.javacourse.seller_rating_app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
-@EnableJpaRepositories("src/main/java/repository")
 public class SellerRatingApplication {
 
 	public static void main(String[] args) {
