+++ src/main/java/com/leverx/javacourse/seller_rating_app/entity/model/Item.java
@@ -1,0 +1,79 @@
+package com.leverx.javacourse.seller_rating_app.entity.model;
+
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Table;
+
+import java.time.LocalDate;
+
+@Entity
+@Table(name = "item")
+public class Item {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private int Id;
+
+    @Column(name = "title")
+    private String title;
+
+    @Column(name = "description")
+    private String description;
+
+    @Column(name = "game_title")
+    private String gameTitle;
+
+    @Column(name = "created")
+    private LocalDate created;
+
+    @Column(name = "updated")
+    private LocalDate updated;
+
+    public Item() {
+    }
+
+    public Item(String title) {
+        this.title = title;
+        this.created = LocalDate.now();
+    }
+
+    public int getId() {
+        return Id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public LocalDate getCreated() {
+        return created;
+    }
+
+    public void setCreated(LocalDate created) {
+        this.created = created;
+    }
+
+    public LocalDate getUpdated() {
+        return updated;
+    }
+
+    public void setUpdated(LocalDate updated) {
+        this.updated = updated;
+    }
+
+    @Override
+    public String toString() {
+        return "Item{" + "title='" + title + '\'' + ", description='" + description + '\'' + ", created=" + created + ", updated=" + updated + '}';
+    }
+}
