+++ src/main/java/com/leverx/javacourse/seller_rating_app/entity/model/UserRoles.java
@@ -1,4 +1,4 @@
-package com.leverx.javacourse.seller_rating_app.entity;
+package com.leverx.javacourse.seller_rating_app.entity.model;
 
 public enum UserRoles {
     ADMINISTRATOR, SELLER, ANONYMOUS_USER
