+++ src/main/java/com/leverx/javacourse/seller_rating_app/repository/UserRepository.java
@@ -1,9 +1,8 @@
 package com.leverx.javacourse.seller_rating_app.repository;
 
-import com.leverx.javacourse.seller_rating_app.entity.User;
+import com.leverx.javacourse.seller_rating_app.entity.model.User;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
 
 import java.math.BigDecimal;
 import java.util.List;
@@ -13,5 +12,5 @@ public interface UserRepository extends CrudRepository<User, Integer> {
     List<User> findByOrderByRatingDesc(BigDecimal rating);
 
     @Query("SELECT u FROM User u WHERE u.rating BETWEEN :ratingStarts AND :ratingEnds ORDER BY u.rating")
-    List<User> findByRatingRange(@Param("ratingStarts") BigDecimal start, @Param("ratingEnds") BigDecimal end);
+    List<User> findByRatingRange(BigDecimal ratingStarts, BigDecimal ratingEnds);
 }
